Move all the unit tests of "mixer" contract to separate file
