+++ contracts/mixer/src/lib.rs
@@ -3,3 +3,6 @@ pub mod state;
 
 #[cfg(test)]
 pub mod test_util;
+
+#[cfg(test)]
+mod tests;
